don't stash the node_modules folder
currently, if you built the web-vault the `node_modules` folder is
stashed which takes a long time. since this is an ignored folder and we
use `npm ci` to build the web-vault we don't really need to stash it.

`npm ci` will always start with a clean build anyway.
see https://docs.npmjs.com/cli/v8/commands/npm-ci